void-docs.{in,1.in}: create script and man page.
This will be shipped in the void-docs package, and can be used to
retrieve documentation pages. It is a template in order to support an
arbitrary installation prefix.

Move man page to section 1, since it is now primarily about the
void-docs utility, but also contains info about the contents of the
void-docs package. It is templated to allow PREFIX substituition.

Use manually written man page instead of converted markdown. - credit to
@flexibeast